SAAS-403 rename UI to UIDemo
