--- packages/platform-ui/src/pages/UI/index.ts
@@ -1,1 +1,0 @@
-export * from './UI';
