+++ packages/platform-ui/src/pages/UIDemo/index.ts
@@ -1,0 +1,1 @@
+export * from './UIDemo';
