+++ packages/platform-ui/src/pages/index.ts
@@ -1,4 +1,4 @@
 export * from './Login';
-export * from './UI';
 export * from './NotFound';
 export * from './Signup';
+export * from './UIDemo';
